Potential fix for pull request finding 'Avoid signed overflow when packing 16-bit FIDs'

+++ src/art.h
@@ -497,7 +497,7 @@ private:
     */
     static constexpr int buildFid(const FloorTileFrmId& floorFrmId, const RoofTileFrmId& roofFrmId)
     {
-        return (floorFrmId.fid() | ((roofFrmId.fid()) << kRoofTileFidShift));
+        return static_cast<int>(static_cast<unsigned int>(floorFrmId.fid()) | (static_cast<unsigned int>(roofFrmId.fid()) << kRoofTileFidShift));
     }
 };
